Refactor evm vanchor leaf fetching to the provider
