+++ package.json
@@ -20,12 +20,12 @@
     "@polkadot/util-crypto": "9.4.1",
     "@polkadot/wasm-crypto": "6.1.1",
     "react-error-overlay": "6.0.9",
-    "@webb-tools/api": "0.1.4-87",
-    "@webb-tools/api-derive": "0.1.4-87",
-    "@webb-tools/types": "0.1.4-87",
-    "@webb-tools/app-util": "0.1.4-87",
-    "@webb-tools/sdk-core": "file:../webb.js/packages/sdk-core/build",
-    "@webb-tools/wasm-utils": "0.1.4-87",
+    "@webb-tools/api": "0.1.4-89",
+    "@webb-tools/api-derive": "0.1.4-89",
+    "@webb-tools/types": "0.1.4-89",
+    "@webb-tools/app-util": "0.1.4-89",
+    "@webb-tools/sdk-core": "0.1.4-89",
+    "@webb-tools/wasm-utils": "0.1.4-89",
     "@webb-tools/contracts": "0.2.2",
     "@webb-tools/interfaces": "0.2.2",
     "@webb-tools/anchors": "0.2.2"
@@ -43,7 +43,6 @@
     "format": "prettier --write \"packages/**/*.{ts,js,jsx,tsx}\""
   },
   "dependencies": {
-    "@material-ui/core": "^4.11.3",
     "@mui/material": "^5.8.7",
     "@mui/styled-engine": "npm:@mui/styled-engine-sc@latest",
     "@mui/styles": "^5.8.7",
@@ -93,14 +92,14 @@
     "@typescript-eslint/eslint-plugin": "^4.14.2",
     "@typescript-eslint/parser": "^4.14.2",
     "@webb-tools/anchors": "0.2.2",
-    "@webb-tools/api": "^0.1.4-87",
-    "@webb-tools/api-derive": "^0.1.4-87",
-    "@webb-tools/app-util": "^0.1.4-87",
-    "@webb-tools/sdk-core": "file:../webb.js/packages/sdk-core/build",
-    "@webb-tools/wasm-utils": "^0.1.4-87",
+    "@webb-tools/api": "0.1.4-89",
+    "@webb-tools/api-derive": "0.1.4-89",
+    "@webb-tools/app-util": "0.1.4-89",
+    "@webb-tools/sdk-core": "0.1.4-89",
+    "@webb-tools/wasm-utils": "0.1.4-89",
     "@webb-tools/contracts": "0.2.2",
     "@webb-tools/interfaces": "0.2.2",
-    "@webb-tools/types": "^0.1.4-87",
+    "@webb-tools/types": "0.1.4-89",
     "babel-plugin-bundled-import-meta": "^0.3.2",
     "cross-env": "^7.0.3",
     "enzyme": "^3.11.0",
