+++ packages/api-providers/package.json
@@ -14,13 +14,13 @@
     "@babel/runtime": "^7.10.2",
     "@polkadot/extension-dapp": "0.44.1",
     "@walletconnect/web3-provider": "^1.7.5",
-    "@webb-tools/sdk-core": "file:../../../webb.js/packages/sdk-core/build",
-    "@webb-tools/app-util": "0.1.4-87",
-    "@webb-tools/api": "0.1.4-87",
-    "@webb-tools/api-derive": "0.1.4-87",
-    "@webb-tools/test-utils": "0.1.4-87",
-    "@webb-tools/types": "0.1.4-87",
-    "@webb-tools/wasm-utils": "0.1.4-87",
+    "@webb-tools/sdk-core": "0.1.4-89",
+    "@webb-tools/app-util": "0.1.4-89",
+    "@webb-tools/api": "0.1.4-89",
+    "@webb-tools/api-derive": "0.1.4-89",
+    "@webb-tools/test-utils": "0.1.4-89",
+    "@webb-tools/types": "0.1.4-89",
+    "@webb-tools/wasm-utils": "0.1.4-89",
     "@webb-tools/anchors": "0.2.2",
     "@webb-tools/contracts": "0.2.2",
     "@webb-tools/interfaces": "0.2.2",
