+++ packages/api-providers/src/web3/relayer-manager.ts
@@ -1,7 +1,7 @@
 // Copyright 2022 @webb-tools/
 // SPDX-License-Identifier: Apache-2.0
 
-import { Note, parseTypedChainId } from '@webb-tools/sdk-core';
+import { MerkleTree, Note, parseTypedChainId } from '@webb-tools/sdk-core';
 import { ethers } from 'ethers';
 
 import {
@@ -14,7 +14,15 @@ import {
 } from '../abstracts';
 import { WebbRelayerManager } from '../abstracts/relayer/webb-relayer-manager';
 import { InternalChainId, typedChainIdToInternalId } from '../chains';
-import { getFixedAnchorAddressForBridge, webbCurrencyIdFromString, WebbError, WebbErrorCodes } from '..';
+import {
+  BridgeStorage,
+  getFixedAnchorAddressForBridge,
+  Storage,
+  VAnchorContract,
+  webbCurrencyIdFromString,
+  WebbError,
+  WebbErrorCodes,
+} from '..';
 
 export class Web3RelayerManager extends WebbRelayerManager {
   async mapRelayerIntoActive(
@@ -189,5 +197,53 @@ export class Web3RelayerManager extends WebbRelayerManager {
       chainId: chainId,
       contractAddress: address,
     });
+  }
+
+  /**
+   * This routine queries the passed relayers for the leaves of an anchor instance on an evm chain.
+   * It validates the leaves with on-chain data, and saves to the storage once validated.
+   * An array of leaves is returned if validated, otherwise null is returned.
+   * @param relayers - A list of relayers
+   * @param contract - A VAnchorContract wrapper for EVM chains.
+   * @param storage - A storage to save the fetched leaves.
+   */
+  async fetchLeavesFromRelayers(
+    relayers: WebbRelayer[],
+    contract: VAnchorContract,
+    storage: Storage<BridgeStorage>
+  ): Promise<string[] | null> {
+    let leaves: string[] = [];
+    const sourceEvmId = await contract.getEvmId();
+
+    // loop through the sourceRelayers to fetch leaves
+    for (let i = 0; i < relayers.length; i++) {
+      const relayerLeaves = await relayers[i].getLeaves(sourceEvmId, contract.inner.address);
+
+      const validLatestLeaf = await contract.leafCreatedAtBlock(
+        relayerLeaves.leaves[relayerLeaves.leaves.length - 1],
+        relayerLeaves.lastQueriedBlock
+      );
+
+      // leaves from relayer somewhat validated, attempt to build the tree
+      if (validLatestLeaf) {
+        // Assume the destination anchor has the same levels as source anchor
+        const levels = await contract.inner.levels();
+        const tree = MerkleTree.createTreeWithRoot(levels, relayerLeaves.leaves, await contract.getLastRoot());
+
+        // If we were able to build the tree, set local storage and break out of the loop
+        if (tree) {
+          leaves = relayerLeaves.leaves;
+
+          await storage.set(contract.inner.address.toLowerCase(), {
+            lastQueriedBlock: relayerLeaves.lastQueriedBlock,
+            leaves: relayerLeaves.leaves,
+          });
+
+          return leaves;
+        }
+      }
+    }
+
+    return null;
   }
 }
