+++ packages/react-environment/package.json
@@ -9,8 +9,8 @@
     "@influxdata/influxdb-client-browser": "^1.8.0",
     "@polkadot/api": "8.7.1",
     "@polkadot/extension-dapp": "0.44.1",
-    "@webb-tools/api": "0.1.4-87",
-    "@webb-tools/sdk-core": "file:../../../webb.js/packages/sdk-core/build",
+    "@webb-tools/api": "0.1.4-89",
+    "@webb-tools/sdk-core": "0.1.4-89",
     "framer-motion": "^6.4.3"
   },
   "devDependencies": {
