+++ packages/vbridge/package.json
@@ -8,10 +8,10 @@
   "dependencies": {
     "@polkadot/api": "8.7.1",
     "@polkadot/ui-keyring": "1.3.1",
-    "@webb-tools/api": "0.1.4-87",
-    "@webb-tools/app-util": "0.1.4-87",
-    "@webb-tools/sdk-core": "file:../../../webb.js/packages/sdk-core/build",
-    "@webb-tools/types": "0.1.4-87",
+    "@webb-tools/api": "0.1.4-89",
+    "@webb-tools/app-util": "0.1.4-89",
+    "@webb-tools/sdk-core": "0.1.4-89",
+    "@webb-tools/types": "0.1.4-89",
     "dayjs": "^1.8.25",
     "i18next": "^19.7.0",
     "react-i18next": "^11.7.1"
